% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{upload_repo}
\alias{upload_repo}
\alias{download_repo}
\title{Upload repo to s3 bucket}
\usage{
upload_repo(dir = ".", paths = c("index.html", "bin", "src"), ...)

download_repo(dir = ".", ...)
}
\arguments{
\item{dir}{Directory of repo}

\item{paths}{Paths (relative to \code{dir}) to include in upload}

\item{...}{Further arguments (such as \code{bucket}, \code{region}, etc.) passed to
methods of \code{\link[aws.s3:s3HTTP]{aws.s3::s3HTTP()}}}
}
\description{
Upload a CRAN-like repository for packages to an AWS S3 bucket.
}
