% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drat.R
\name{init_repo}
\alias{init_repo}
\alias{insert_pkg}
\alias{insert_pkgs}
\title{Initialize and populate a drat repository}
\usage{
init_repo(dir = ".")

insert_pkg(pkg, dir = ".", ...)

insert_pkgs(pkgs, ...)
}
\arguments{
\item{dir}{Repository directory}

\item{pkg, pkgs}{R package bundle(s)}

\item{...}{Passed on to \link[tools:writePACKAGES]{tools::write_PACKAGES}}
}
\description{
By creating the \code{src/contrib} path and a top-level \code{index.html} file,
\code{init_repo()} initializes a drat repository. Following that, packages can
be added using \code{insert_pkg()}/\code{insert_pkgs()}.
}
\author{
Dirk Eddelbuettel and Nicolas Bennett
}
